Cows and Bulls for Python
Inspired by Exercise 18 from https://www.practicepython.org/exercise/2014/07/05/18-cows-and-bulls.html

The exercise did not have a satisfying solution, so I wrote one that works properly. My script will avoid counting duplicate numbers as bulls as is the case with the initial solution.

To clarify: Say the randomly generated number is 4314. When the user guess is 4444, the program will answer with 2 cows and 0 bulls. Alternatively if the user input is 4154, it will give you the answer: 2 cow, 1 bull.